backend/requirements.txt
pyserial==3.5
flask==2.3.3
flask-cors==4.0.0